add sample data
test sample data

move data outside of package folder

compress gz; add i2i test

bugfix

update README
