+++ .github/workflows/python-app.yml
@@ -54,6 +54,9 @@ jobs:
     - name: Pip freeze
       run: |
         pip freeze
+    - name: Download test data
+      run: |
+        (cd data/amazon_review_prime_pantry && wget http://deepyeti.ucsd.edu/jianmo/amazon/metaFiles2/meta_Prime_Pantry.json.gz)
     - name: Test with pytest
       run: |
         pytest
