+++ setup.cfg
@@ -1,0 +1,7 @@
+[tool:pytest]
+norecursedirs = data logs .git
+testpaths = test
+addopts =
+    --verbose
+    --durations=5
+    --color=yes
