+++ test/test_prime_pantry.py
@@ -5,85 +5,25 @@ from ccrec import InteractiveExperiment, env
 from ccrec.env import create_zero_shot, Env, parse_response
 from ccrec.env.i2i_env import get_notebook_name
 from ccrec.models.vae_models import VAEPretrainedModel
+from ccrec.util.amazon_review_prime_pantry import get_item_df
 
 
-def _shorten_brand_name_function(x):
-    if not isinstance(x, str):
-        return x
-    for i, part in enumerate(x.split(' ')):
-        try:
-            int(part)
-            continue
-        except Exception:
-            return ' '.join(x.split(' ')[:i + 1])
+def test_sample_data():
+    return get_item_df()
 
 
-def get_item_df(meta_file='data/prime_pantry/pp-items-full-meta.csv',
-                landingImageURL_file='data/prime_pantry/landingImageURL-new.csv',
-                landingImageURL_folder='data/prime_pantry/landingImage',
-                shorten_brand_name=True,
-                return_tfidf_csr=True):
-    item_df = pd.read_csv(meta_file)
-    item_df = item_df.set_index('ITEM_ID').assign(
-        TITLE=lambda df: df['DESCRIPTION'].apply(lambda x: ' '.join(x.split(' ')[:30]))
-    )[['TITLE', 'BRAND']]
-
-    if shorten_brand_name:
-        item_df['BRAND'] = item_df['BRAND'].apply(_shorten_brand_name_function)
-    print(f'# items {len(item_df)}, # brands {item_df["BRAND"].nunique()}')
-
-    item_df = item_df.join(
-        pd.read_csv(landingImageURL_file, names=['asin', 'landingImage']).set_index('asin')
-    )
-    item_df = item_df[item_df['landingImage'].notnull()]
-    item_df = item_df[item_df['landingImage'].apply(lambda x: x.endswith('.jpg'))]
-    item_df['landingImage'] = [f'{landingImageURL_folder}/{x}.jpg' for x in item_df.index.values]
-
-    tfidf_fit = TfidfVectorizer().fit(item_df['TITLE'].tolist())
-    tfidf_csr = tfidf_fit.transform(item_df['TITLE'].tolist())
-
-    item_df['tfidf_indices'] = np.split(tfidf_csr.indices, tfidf_csr.indptr[1:-1])
-    item_df['tfidf_words'] = np.split(
-        np.array(tfidf_fit.get_feature_names())[tfidf_csr.indices],
-        tfidf_csr.indptr[1:-1])
-    item_df['tfidf_data'] = np.split(tfidf_csr.data, tfidf_csr.indptr[1:-1])
-
-    item_df = item_df.join(
-        pd.concat({
-            'words': item_df['tfidf_words'].explode(),
-            'data': item_df['tfidf_data'].explode(),
-        }, axis=1).sort_values('data', ascending=False)
-        .groupby('ITEM_ID')['words'].apply(lambda x: x[:5].tolist())
-        .to_frame('sorted_words')
-    ).drop('tfidf_words', axis=1)
-
-    return (item_df, tfidf_csr) if return_tfidf_csr else item_df
-
-
-def get_interactions_df(interactions_file='data/prime_pantry/pp-all_interactions.csv'):
-    return pd.read_csv(interactions_file).sort_values("TIMESTAMP", kind='mergesort')
-
-
-def get_collection_event_df(event_file='prime-pantry-i2i-online-baseline4.json'):
-    return pd.read_json(event_file, lines=True, convert_dates=False)
-
-
-def get_collection_response(response_file='prime-pantry-i2i-online-baseline4-response.json'):
-    response = pd.read_json(response_file, lines=True, convert_dates=False)
-    return response.set_index(response.columns[:2].tolist())
-
-
-@pytest.mark.skip(reason='requires data')
 def test_prime_pantry_ccrec(
     simulation=True,
     pretrained_checkpoint=None,
     train_requests=None,  # subsample training queries
     epsilon=0,
-    max_epochs=1,  # choose 0 to skip retraining
+    max_epochs=0,  # choose 0 to skip retraining
     role_arn=None,
     s3_prefix=None,
     working_model=None,  # VAEPretrainedModel
     multi_label=False,
+    gnd_response_json='data/amazon_review_prime_pantry/prime_pantry_test_response.json.gz',
+    n_steps=1,
 ):
     """
     * pretrained_checkpoint
@@ -101,7 +41,7 @@ def test_prime_pantry_ccrec(
     user_df = zero_shot.user_df
 
     gnd_response = pd.read_json(
-        'prime-pantry-i2i-online-baseline4-response.json', lines=True, convert_dates=False,
+        gnd_response_json, lines=True, convert_dates=False,
     ).rename({'level_0': 'USER_ID'}, axis=1).set_index(['USER_ID', 'TEST_START_TIME'])
     gnd_events = pd.concat([
         zero_shot.event_df,  # history features
@@ -110,7 +50,7 @@ def test_prime_pantry_ccrec(
 
     if working_model is None:
         working_model = ccrec.models.bbpr.BertBPR(
-            item_df, max_epochs=max_epochs, batch_size=10 * torch.cuda.device_count(),
+            item_df, max_epochs=max_epochs, batch_size=10 * max(1, torch.cuda.device_count()),
             sample_with_prior=True, sample_with_posterior=0, elementwise_affine=False,
             replacement=False, n_negatives=5, valid_n_negatives=5,
             training_prior_fcn=lambda x: (x + 1 / x.shape[1]).clip(0, None).log(),
@@ -153,7 +93,7 @@ def test_prime_pantry_ccrec(
         working_model, baseline_models, epsilon,
     )
 
-    iexp.run(n_steps=10, test_every=1, test_before_train=True)
+    iexp.run(n_steps=n_steps, test_every=1, test_before_train=True)
     print(pd.DataFrame(iexp.testing_env._reward_by_policy))
 
     return iexp
